refactor: improve tencent sign
